refactor(navbar): move to script setup
