+++ src/config.json
@@ -210,6 +210,7 @@
           "name": "Navbar",
           "cName": "头部导航",
           "desc": "头部导航",
+          "setup": true,
           "author": "liqiong43"
         },
         {
