--- src/packages/__VUE/navbar/index.vue
@@ -1,142 +1,0 @@
-<template>
-  <view class="nut-navbar--placeholder" :style="rootStyle">
-    <view ref="navbarRef" :class="classes" :style="{ zIndex }">
-      <view class="nut-navbar__left" @click="handleLeft">
-        <slot v-if="leftShow" name="left-show">
-          <Left height="12px" color="#979797"></Left>
-        </slot>
-        <view v-if="leftText" class="nut-navbar__text">{{ leftText }}</view>
-        <slot name="left"></slot>
-      </view>
-      <view class="nut-navbar__title">
-        <view v-if="title" class="title" @click="handleCenter">{{ title }}</view>
-        <view v-if="titleIcon" class="icon" @click="handleCenterIcon">
-          <slot name="title-icon" @click="handleCenterIcon"></slot>
-        </view>
-        <slot name="content"></slot>
-      </view>
-      <view class="nut-navbar__right" @click="handleRight">
-        <view v-if="desc" class="nut-navbar__text">{{ desc }}</view>
-        <slot name="right"></slot>
-      </view>
-    </view>
-  </view>
-</template>
-
-<script lang="ts">
-import { onMounted, computed, toRefs, ref, nextTick } from 'vue';
-import { createComponent } from '@/packages/utils/create';
-import { Left } from '@nutui/icons-vue';
-const { componentName, create } = createComponent('navbar');
-export default create({
-  components: { Left },
-  props: {
-    leftShow: { type: Boolean, default: false }, //左侧  是否显示返回icon
-    title: { type: String, default: '' }, //中间  文字标题
-    titleIcon: { type: Boolean, default: false }, //中间
-    leftText: { type: String, default: '' }, //左侧文字
-    desc: { type: String, default: '' }, //右侧   按钮文字
-    fixed: {
-      type: Boolean,
-      default: false
-    },
-    safeAreaInsetTop: {
-      type: Boolean,
-      default: false
-    },
-    border: {
-      type: Boolean,
-      default: false
-    },
-    placeholder: {
-      // 生成一个等高的占位元素
-      type: Boolean,
-      default: true
-    },
-    zIndex: {
-      type: [Number, String],
-      default: 10
-    }
-  },
-  emits: [
-    'clickBack',
-    'clickTitle',
-    'clickIcon',
-    'clickRight',
-    // will be deprecated
-    'onClickBack',
-    'onClickTitle',
-    'onClickIcon',
-    'onClickRight'
-  ],
-  setup(props, { emit }) {
-    const { border, fixed, safeAreaInsetTop, placeholder } = toRefs(props);
-    const navHeight = ref('auto');
-    const navbarRef = ref<HTMLElement>();
-    const classes = computed(() => {
-      const prefixCls = componentName;
-      return {
-        [prefixCls]: true,
-        [`${prefixCls}--border`]: border.value,
-        [`${prefixCls}--fixed`]: fixed.value,
-        [`${prefixCls}--safe-area-inset-top`]: safeAreaInsetTop.value
-      };
-    });
-
-    const rootStyle = computed(() => {
-      if (fixed.value && placeholder.value) {
-        return {
-          height: navHeight.value
-        };
-      }
-      return {};
-    });
-
-    const getNavHeight = () => {
-      if (navbarRef.value) {
-        const rect = navbarRef.value.getBoundingClientRect();
-        navHeight.value = `${rect.height}px`;
-      }
-    };
-
-    onMounted(() => {
-      if (fixed.value && placeholder.value) {
-        nextTick(() => {
-          getNavHeight();
-        });
-      }
-    });
-
-    const handleLeft = () => {
-      emit('clickBack');
-      emit('onClickBack');
-    };
-
-    const handleCenter = () => {
-      emit('clickTitle');
-      emit('onClickTitle');
-    };
-
-    const handleCenterIcon = () => {
-      emit('clickIcon');
-      emit('onClickIcon');
-    };
-
-    const handleRight = () => {
-      emit('clickRight');
-      emit('onClickRight');
-    };
-
-    return {
-      navbarRef,
-      rootStyle,
-      classes,
-      navHeight,
-      handleLeft,
-      handleCenter,
-      handleCenterIcon,
-      handleRight
-    };
-  }
-});
-</script>
